requirements.txt
appnope==0.1.2
asgiref==3.4.1
backcall==0.2.0
boto3==1.18.6
botocore==1.21.6
certifi==2021.5.30
charset-normalizer==2.0.3
decorator==5.0.9
Django==3.2.5
django-cors-headers==3.7.0
django-extensions==3.1.3
django-request==1.5.6
django-storages==1.11.1
idna==3.2
ihttp==0.3
ipython==7.25.0
ipython-genutils==0.2.0
jedi==0.18.0
jmespath==0.10.0
matplotlib-inline==0.1.2
mysqlclient==2.0.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.19
ptyprocess==0.7.0
Pygments==2.9.0
PyJWT==2.1.0
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.1
traitlets==5.0.5
urllib3==1.26.6
wcwidth==0.2.5
